+++ Swift/OmFileFormat/FileExtension.swift
@@ -1,5 +1,6 @@
 import Foundation
 
+
 extension FileHandle {
     /// Create new file and convert it into a `FileHandle`. For some reason this does not exist in stock swift....
     /// Error on existing file
